server/requirements.txt
# Flask web framework
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0

# Database
pymongo==4.13.2

# LangChain - Python 3.13 compatible versions
langchain==0.3.9
langsmith==0.1.129
langchain-openai==0.2.9
langchain-community==0.3.9

# AI/ML - Updated OpenAI version to match langchain-openai requirements
openai>=1.54.0

# Vector database
pinecone-client==5.0.1

# Hugging Face datasets - Python 3.13 compatible
datasets==3.2.0
transformers==4.48.0

# Environment and utilities
python-dotenv==1.1.0
requests==2.32.4

# Development tools
pylint==3.3.7

# Core dependencies - Python 3.13 compatible versions
numpy>=2.0.0
certifi==2025.6.15
urllib3==2.4.0
click==8.2.1
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
itsdangerous==2.2.0
blinker==1.9.0

# Async support
aiohttp==3.12.13
httpx==0.28.1

# Additional dependencies that might be needed
packaging>=23.0
typing-extensions>=4.0.0
pydantic>=2.0.0